mt_evtCustomEvent: FUN_0039e1b0 fix two real bugs. (1) uVar4 declared u64 but assigned from FUN_00386ae0 which returns plain int (per its prototype in mt_evtMisc.c; no prototype is visible in this TU so MWCC implicit-declares it int-returning) - the u64 declaration forced a spurious 32->64 sign-extension widening (dsll32/dsra32) at every one of its 12 case-body assignments; retyped to u32 to match the real return width. (2) switch(*(u8*)(param_2+0x10)) used an unsigned byte load (lbu) where retail uses a signed byte load (lb) for the same switch selector - retyped the cast to s8. diff 844->808, object_size 1340->1328. Residual is a distinct register-coloring difference (param_1/param_2 held in swapped -registers vs retail) compounded with raw-hex-literal-vs-symbol codegen (0x6a1660 etc. compile via lui+ori instead of retail's lui+addiu) - tested converting the literals to DAT_ symbols in isolation and confirmed relocation names now match retail exactly, but it interacts with the still-unresolved register swap and made the aggregate diff worse (808->892); reverted that specific sub-fix pending resolution of the register-coloring issue first.

+++ src/Event/mt_evtCustomEvent.c
@@ -5429,7 +5429,7 @@ u32 FUN_0039e1b0(int param_1,int param_2)
 
   u32 uVar3;
 
-  u64 uVar4;
+  u32 uVar4;
 
   
 
@@ -5441,7 +5441,7 @@ u32 FUN_0039e1b0(int param_1,int param_2)
 
   }
 
-  switch(*(u8 *)(param_2 + 0x10)) {
+  switch(*(s8 *)(param_2 + 0x10)) {
 
   case 0:
 
@@ -5458,7 +5458,6 @@ u32 FUN_0039e1b0(int param_1,int param_2)
     uVar3 = FUN_00194b20(*(u32 *)(param_1 + 0x92c),0x6a1660,0x10,0x398530,0x399830,uVar4);
 
     *(u32 *)(param_1 + 0x938) = uVar3;
-
     break;
 
   case 1:
